Add Security/Open rubocop rule
The `open` method can be used to open files or URLs and it's deprecated
in Ruby 2.7. In this case, it's clear we're dealing with a URL, so we
can use `URI.open`.

The code is a bit strange, though, since it returns a value and it has a
side effect, opening the URL. I'm not sure about the intention of the
code; my best guess is we want to test the URL exists and is accessible
before returning it (and, if that's the case, IMHO the code should be a
bit more explicit in order to show the intention behind it), but it
could also be an unintended side effect which is there by accident.